added an env bypass
