+++ config.js
@@ -104,6 +104,8 @@ module.exports =
     grants: process.env.DISABLE_GRANTS || false
   },
   forward: {
+    settlements: process.env.FORWARD_SETTLEMENTS_TO_KAFKA || false,
+    referrals: process.env.FORWARD_REFERRALS_TO_KAFKA || false,
     wallets: process.env.FORWARD_WALLET_TO_GRANTS || false,
     grants: process.env.FORWARD_TO_GRANTS || false
   },
